backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
PyJWT==2.8.0
psycopg2-binary==2.9.7
bcrypt==4.0.1
gunicorn==21.2.0
prometheus-flask-exporter==0.22.4